+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+- `getBrightnessLevel` and `getVolumeLevel` now return `null`.
+- `startHeadTracking` now returns `Map<String, dynamic>`.
+
 ## 0.1.1
 - Windows support
 - Android support